+++ src/sysdolphin/baselib/hsd_3933.c
@@ -349,7 +349,7 @@ int hsd_80393A5C(char* filename, int data, int size)
     }
 
     fd_arg = fd;
-    data_p = (u32*) data;
+    data_p = (u32*) (uintptr_t) data;
     written_f = (f32) FIOFwrite(fd_arg, data_p, size);
 
     if ((f32) (s32) size != written_f) {
